For visualizing the 3'bias in rna seq data, it is important to plot the reads coverage over gene body based on the relative distance to the polyA tail that is measured over the mNRA transcript, instead of measured over the genome assembly. It is also important to maintain the actuall basepair distance in the counting by cuting the longer genes and omitting the shorter genes. Instead of shinking longer genes and stretching shorter genes to a standard length.
To achieve these goals, a new cython class:
"GeneCounterAbsoluteDistanceFromThreePrimeEnd" is created. This class
inherits the class "IntervalsCounter". The derived class (1) overrides
the "__init__" of the baseclass; (2) implements a private method
"__chopAllTranscriptToAFixedLengthForTheThreePrimeBiasCounting" to
standardize all genes into a fixed length; (3) and implements the public
method "count" that is the starndard interface to performs the actual
counting similiar to all other counting classes in this script.

Additional remarks:

(*) Correpsonding test cases for this new feature are included in the
same commit transaction.

(*) Minor edits to other parts of the scripts and existing test cases.

(*) "geneprofilewithintrons" and
"geneprofileabsolutedistancefromthreeprimeend" always imply
--base-accuracy.

(*) Fully backward compatible.