1.Add one recipy to match CGAT paper. 2.Fix two other doc discrepancies.
Details: 1.Figure 1(a) in CGAT paper use geneprofile, with default
parameters. Thus I added this to the recipies in the script
documentation.
2.The documentation incorrectly descibed using "--counter" to select
different gene models, instead of using "--method".
The documentation incorrectly described the base accurary option is
turned on by default, which then can be turn off by specifying
"--base-accuracy-off". Instead, the base accurary option is off by
default, and needs to be turned on by specifying "--base-accuracy".